web/src/requirements.txt
pyramid==2.0
pyramid-jinja2
python-dotenv
waitress
pandas
numpy
matplotlib
scikit-learn==0.24.1
mysql-connector-python

